refactor: introduce @bitwarden/device-type
