+++ libs/device-type/README.md
@@ -1,0 +1,5 @@
+# device-type
+
+Owned by: platform
+
+Constants and helpers for working with specific device types
