+++ libs/device-type/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "@bitwarden/device-type",
+  "version": "0.0.1",
+  "description": "Constants and helpers for working with specific device types",
+  "private": true,
+  "type": "commonjs",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "license": "GPL-3.0",
+  "author": "platform"
+}
