+++ libs/device-type/src/device-type.spec.ts
@@ -1,0 +1,8 @@
+import * as lib from "./index";
+
+describe("device-type", () => {
+  // This test will fail until something is exported from index.ts
+  it("should work", () => {
+    expect(lib).toBeDefined();
+  });
+});
