+++ tsconfig.base.json
@@ -29,6 +29,7 @@
       "@bitwarden/common/*": ["./libs/common/src/*"],
       "@bitwarden/components": ["./libs/components/src"],
       "@bitwarden/core-test-utils": ["libs/core-test-utils/src/index.ts"],
+      "@bitwarden/device-type": ["libs/device-type/src/index.ts"],
       "@bitwarden/dirt-card": ["./libs/dirt/card/src"],
       "@bitwarden/generator-components": ["./libs/tools/generator/components/src"],
       "@bitwarden/generator-core": ["./libs/tools/generator/core/src"],
